requirements.txt
alembic==1.4.3
blinker==1.4
Bootstrap-Flask==1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
flask-whooshalchemy3==0.2.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pbr==3.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
Whoosh==2.7.4
WTForms==2.3.3
